[UPDATE] separate base functionality
