+++ rviz_default_plugins/src/rviz_default_plugins/displays/map/map_display.cpp
@@ -30,7 +30,6 @@
 
 #include "rviz_default_plugins/displays/map/map_display.hpp"
 
-#include <algorithm>
 #include <memory>
 #include <string>
 #include <vector>
@@ -39,111 +38,18 @@
 #include <OgreSceneNode.h>
 #include <OgreTextureManager.h>
 #include <OgreTechnique.h>
-#include <OgreSharedPtr.h>
 
-#include "rclcpp/time.hpp"
 
-#include "rviz_rendering/material_manager.hpp"
 #include "rviz_rendering/objects/grid.hpp"
-#include "rviz_common/logging.hpp"
-#include "rviz_common/msg_conversions.hpp"
-#include "rviz_common/properties/enum_property.hpp"
-#include "rviz_common/properties/float_property.hpp"
-#include "rviz_common/properties/int_property.hpp"
 #include "rviz_common/properties/property.hpp"
-#include "rviz_common/properties/quaternion_property.hpp"
-#include "rviz_common/properties/ros_topic_property.hpp"
-#include "rviz_common/properties/vector_property.hpp"
 #include "rviz_common/validate_floats.hpp"
 #include "rviz_common/display_context.hpp"
-#include "rviz_default_plugins/displays/map/palette_builder.hpp"
 
 
 namespace rviz_default_plugins
 {
 namespace displays
 {
-
-MapDisplay::MapDisplay()
-: loaded_(false),
-  resolution_(0.0f),
-  width_(0),
-  height_(0),
-  update_profile_(rclcpp::QoS(5)),
-  update_messages_received_(0)
-{
-  connect(this, SIGNAL(mapUpdated()), this, SLOT(showMap()));
-
-  update_topic_property_ = new rviz_common::properties::RosTopicProperty(
-    "Update Topic", "",
-    "", "Topic where updates to this map display are received. "
-    "This topic is automatically determined by the map topic. "
-    "If the map is received on 'map_topic', the display assumes updates are received on "
-    "'map_topic_updates'."
-    "This can be overridden in the UI by clicking on the topic and setting the desired topic.",
-    this, SLOT(updateMapUpdateTopic()));
-
-  update_profile_property_ = new rviz_common::properties::QosProfileProperty(
-    update_topic_property_, update_profile_);
-
-  alpha_property_ = new rviz_common::properties::FloatProperty(
-    "Alpha", 0.7f,
-    "Amount of transparency to apply to the map.",
-    this, SLOT(updateAlpha()));
-  alpha_property_->setMin(0);
-  alpha_property_->setMax(1);
-
-  color_scheme_property_ = new rviz_common::properties::EnumProperty(
-    "Color Scheme", "map",
-    "How to color the occupancy values.",
-    this, SLOT(updatePalette()));
-  // Option values here must correspond to indices in palette_textures_ array in onInitialize()
-  // below.
-  color_scheme_property_->addOption("map", 0);
-  color_scheme_property_->addOption("costmap", 1);
-  color_scheme_property_->addOption("raw", 2);
-
-  draw_under_property_ = new rviz_common::properties::BoolProperty(
-    "Draw Behind", false,
-    "Rendering option, controls whether or not the map is always"
-    " drawn behind everything else.",
-    this, SLOT(updateDrawUnder()));
-
-  resolution_property_ = new rviz_common::properties::FloatProperty(
-    "Resolution", 0,
-    "Resolution of the map. (not editable)", this);
-  resolution_property_->setReadOnly(true);
-
-  width_property_ = new rviz_common::properties::IntProperty(
-    "Width", 0,
-    "Width of the map, in meters. (not editable)", this);
-  width_property_->setReadOnly(true);
-
-  height_property_ = new rviz_common::properties::IntProperty(
-    "Height", 0,
-    "Height of the map, in meters. (not editable)", this);
-  height_property_->setReadOnly(true);
-
-  position_property_ = new rviz_common::properties::VectorProperty(
-    "Position", Ogre::Vector3::ZERO,
-    "Position of the bottom left corner of the map, in meters. (not editable)",
-    this);
-  position_property_->setReadOnly(true);
-
-  orientation_property_ = new rviz_common::properties::QuaternionProperty(
-    "Orientation", Ogre::Quaternion::IDENTITY, "Orientation of the map. (not editable)", this);
-  orientation_property_->setReadOnly(true);
-
-  transform_timestamp_property_ = new rviz_common::properties::BoolProperty(
-    "Use Timestamp", false,
-    "Use map header timestamp when transforming", this, SLOT(transformMap()));
-}
-
-MapDisplay::~MapDisplay()
-{
-  unsubscribe();
-  clear();
-}
 
 Ogre::TexturePtr makePaletteTexture(std::vector<unsigned char> palette_bytes)
 {
@@ -157,31 +63,8 @@ Ogre::TexturePtr makePaletteTexture(std::vector<unsigned char> palette_bytes)
 }
 
 MapDisplay::MapDisplay(rviz_common::DisplayContext * context)
-: MapDisplay()
+: MapDisplayBase<nav_msgs::msg::OccupancyGrid, map_msgs::msg::OccupancyGridUpdate>(context)
 {
-  context_ = context;
-  scene_manager_ = context->getSceneManager();
-  scene_node_ = scene_manager_->getRootSceneNode()->createChildSceneNode();
-
-  palette_textures_.push_back(makePaletteTexture(makeMapPalette()));
-  color_scheme_transparency_.push_back(false);
-  palette_textures_.push_back(makePaletteTexture(makeCostmapPalette()));
-  color_scheme_transparency_.push_back(true);
-  palette_textures_.push_back(makePaletteTexture(makeRawPalette()));
-  color_scheme_transparency_.push_back(true);
-}
-
-void MapDisplay::onInitialize()
-{
-  MFDClass::onInitialize();
-  rviz_ros_node_ = context_->getRosNodeAbstraction();
-  update_topic_property_->initialize(rviz_ros_node_);
-
-  update_profile_property_->initialize(
-    [this](rclcpp::QoS profile) {
-      this->update_profile_ = profile;
-      updateMapUpdateTopic();
-    });
   // Order of palette textures here must match option indices for color_scheme_property_ above.
   palette_textures_.push_back(makePaletteTexture(makeMapPalette()));
   color_scheme_transparency_.push_back(false);
@@ -189,135 +72,28 @@ void MapDisplay::onInitialize()
   color_scheme_transparency_.push_back(true);
   palette_textures_.push_back(makePaletteTexture(makeRawPalette()));
   color_scheme_transparency_.push_back(true);
+
+  // add property for color_map selection
+  color_scheme_property_ = new rviz_common::properties::EnumProperty(
+    "Color Scheme", "map",
+    "How to color the occupancy values.",
+    this, SLOT(updatePalette()));
+  // Option values here must correspond to indices in palette_textures_ array above
+  color_scheme_property_->addOption("map", 0);
+  color_scheme_property_->addOption("costmap", 1);
+  color_scheme_property_->addOption("raw", 2);
 }
 
-void MapDisplay::updateTopic()
+void MapDisplay::showValidMap()
 {
-  update_topic_property_->setValue(topic_property_->getTopic() + "_updates");
-  MFDClass::updateTopic();
+  MapDisplayBase::showValidMap();
+  updatePalette();
 }
 
-void MapDisplay::subscribe()
-{
-  if (!isEnabled()) {
-    return;
-  }
-
-  if (topic_property_->isEmpty()) {
-    setStatus(
-      rviz_common::properties::StatusProperty::Error,
-      "Topic",
-      QString("Error subscribing: Empty topic name"));
-    return;
-  }
-
-  MFDClass::subscribe();
-
-  subscribeToUpdateTopic();
-}
-
-void MapDisplay::subscribeToUpdateTopic()
-{
-  try {
-    rclcpp::SubscriptionOptions sub_opts;
-    sub_opts.event_callbacks.message_lost_callback =
-      [&](rclcpp::QOSMessageLostInfo & info)
-      {
-        std::ostringstream sstm;
-        sstm << "Some messages were lost:\n>\tNumber of new lost messages: " <<
-          info.total_count_change << " \n>\tTotal number of messages lost: " <<
-          info.total_count;
-        setStatus(
-          rviz_common::properties::StatusProperty::Warn, "Update Topic",
-          QString(sstm.str().c_str()));
-      };
-
-    update_subscription_ =
-      rviz_ros_node_.lock()->get_raw_node()->
-      template create_subscription<map_msgs::msg::OccupancyGridUpdate>(
-      update_topic_property_->getTopicStd(), update_profile_,
-      [this](const map_msgs::msg::OccupancyGridUpdate::ConstSharedPtr message) {
-        incomingUpdate(message);
-      },
-      sub_opts);
-    setStatus(rviz_common::properties::StatusProperty::Ok, "Update Topic", "OK");
-  } catch (rclcpp::exceptions::InvalidTopicNameError & e) {
-    setStatus(
-      rviz_common::properties::StatusProperty::Error, "Update Topic",
-      QString("Error subscribing: ") + e.what());
-  }
-}
-
-void MapDisplay::unsubscribe()
-{
-  MFDClass::unsubscribe();
-  unsubscribeToUpdateTopic();
-}
-
-void MapDisplay::unsubscribeToUpdateTopic()
-{
-  update_subscription_.reset();
-}
-
-void MapDisplay::updateAlpha()
-{
-  float alpha = alpha_property_->getFloat();
-  Ogre::SceneBlendType scene_blending;
-  bool depth_write;
-
-  rviz_rendering::MaterialManager::enableAlphaBlending(scene_blending, depth_write, alpha);
-
-  for (const auto & swatch : swatches_) {
-    swatch->updateAlpha(scene_blending, depth_write, alpha);
-  }
-}
-
-void MapDisplay::updateDrawUnder() const
-{
-  bool draw_under = draw_under_property_->getValue().toBool();
-
-  if (alpha_property_->getFloat() >= rviz_rendering::unit_alpha_threshold) {
-    for (const auto & swatch : swatches_) {
-      swatch->setDepthWriteEnabled(!draw_under);
-    }
-  }
-
-  uint8_t group = draw_under ? Ogre::RENDER_QUEUE_4 : Ogre::RENDER_QUEUE_MAIN;
-  for (const auto & swatch : swatches_) {
-    swatch->setRenderQueueGroup(group);
-  }
-}
-
-void MapDisplay::clear()
-{
-  if (isEnabled()) {
-    setStatus(rviz_common::properties::StatusProperty::Warn, "Message", "No map received");
-  }
-
-  if (!loaded_) {
-    return;
-  }
-
-  swatches_.clear();
-  height_ = 0;
-  width_ = 0;
-  resolution_ = 0.0f;
-
-  loaded_ = false;
-}
-
-bool validateFloats(const nav_msgs::msg::OccupancyGrid & msg)
+bool MapDisplay::validateFloats(const nav_msgs::msg::OccupancyGrid & msg) const
 {
   return rviz_common::validateFloats(msg.info.resolution) &&
          rviz_common::validateFloats(msg.info.origin);
-}
-
-void MapDisplay::processMessage(nav_msgs::msg::OccupancyGrid::ConstSharedPtr msg)
-{
-  current_map_ = *msg;
-  loaded_ = true;
-  // updated via signal in case ros spinner is in a different thread
-  Q_EMIT mapUpdated();
 }
 
 void MapDisplay::incomingUpdate(const map_msgs::msg::OccupancyGridUpdate::ConstSharedPtr update)
@@ -344,7 +120,7 @@ void MapDisplay::incomingUpdate(const map_msgs::msg::OccupancyGridUpdate::ConstS
   setStatus(rviz_common::properties::StatusProperty::Ok, "Update", "Update OK");
 
   // updated via signal in case ros spinner is in a different thread
-  Q_EMIT mapUpdated();
+  Q_EMIT q_helper_object_->mapUpdated();
 }
 
 bool MapDisplay::updateDataOutOfBounds(
@@ -367,205 +143,12 @@ void MapDisplay::updateMapDataInMemory(
   }
 }
 
-void MapDisplay::createSwatches()
+void MapDisplay::processMessage(nav_msgs::msg::OccupancyGrid::ConstSharedPtr msg)
 {
-  size_t width = current_map_.info.width;
-  size_t height = current_map_.info.height;
-  float resolution = current_map_.info.resolution;
-
-  size_t swatch_width = width;
-  size_t swatch_height = height;
-  int number_swatches = 1;
-  // One swatch can have up to 2^16 * 2^16 pixel (8 bit texture, i.e. 4GB of data)
-  // Since the width and height are separately limited by 2^16 it might be necessary to have several
-  // pieces, however more than 8 swatches is probably unnecessary due to memory limitations
-  const size_t maximum_number_swatch_splittings = 4;
-
-  for (size_t i = 0; i < maximum_number_swatch_splittings; ++i) {
-    RVIZ_COMMON_LOG_INFO_STREAM(
-      "Trying to create a map of size " <<
-        width << " x " << height << " using " << number_swatches << " swatches");
-    swatches_.clear();
-    try {
-      tryCreateSwatches(width, height, resolution, swatch_width, swatch_height, number_swatches);
-      updateDrawUnder();
-      return;
-    } catch (Ogre::InvalidParametersException &) {
-      doubleSwatchNumber(swatch_width, swatch_height, number_swatches);
-    } catch (Ogre::RenderingAPIException &) {
-      // This exception seems no longer thrown on some systems. May still be relevant for others.
-      doubleSwatchNumber(swatch_width, swatch_height, number_swatches);
-    }
-  }
-  RVIZ_COMMON_LOG_ERROR_STREAM(
-    "Creating " << number_swatches << "failed. "
-      "This map is too large to be displayed by RViz.");
-  swatches_.clear();
-}
-
-void MapDisplay::doubleSwatchNumber(
-  size_t & swatch_width, size_t & swatch_height, int & number_swatches) const
-{
-  RVIZ_COMMON_LOG_ERROR_STREAM(
-    "Failed to create map using " << number_swatches << " swatches. "
-      "At least one swatch seems to need too much memory");
-  if (swatch_width > swatch_height) {
-    swatch_width /= 2;
-  } else {
-    swatch_height /= 2;
-  }
-  number_swatches *= 2;
-}
-
-void MapDisplay::tryCreateSwatches(
-  size_t width,
-  size_t height,
-  float resolution,
-  size_t swatch_width,
-  size_t swatch_height,
-  int number_swatches)
-{
-  size_t x = 0;
-  size_t y = 0;
-  for (int i = 0; i < number_swatches; i++) {
-    size_t effective_width = getEffectiveDimension(width, swatch_width, x);
-    size_t effective_height = getEffectiveDimension(height, swatch_height, y);
-
-    swatches_.push_back(
-      std::make_shared<Swatch>(
-        scene_manager_,
-        scene_node_,
-        x, y,
-        effective_width,
-        effective_height,
-        resolution,
-        draw_under_property_->getValue().toBool()));
-
-    swatches_[i]->updateData(current_map_);
-
-    x += effective_width;
-    if (x >= width) {
-      x = 0;
-      y += effective_height;
-    }
-  }
-  updateAlpha();
-}
-
-size_t MapDisplay::getEffectiveDimension(
-  size_t map_dimension, size_t swatch_dimension, size_t position)
-{
-  // Last swatch is bigger than swatch_dimension for odd numbers.
-  // subtracting the swatch_dimension in the LHS handles this case.
-  return map_dimension - position - swatch_dimension >= swatch_dimension ?
-         swatch_dimension :
-         map_dimension - position;
-}
-
-void MapDisplay::showMap()
-{
-  if (current_map_.data.empty()) {
-    return;
-  }
-
-  if (!validateFloats(current_map_)) {
-    setStatus(
-      rviz_common::properties::StatusProperty::Error, "Map",
-      "Message contained invalid floating point values (nans or infs)");
-    return;
-  }
-
-  size_t width = current_map_.info.width;
-  size_t height = current_map_.info.height;
-
-  if (width * height == 0) {
-    std::string message =
-      "Map is zero-sized (" + std::to_string(width) + "x" + std::to_string(height) + ")";
-    setStatus(
-      rviz_common::properties::StatusProperty::Error, "Map", QString::fromStdString(message));
-    return;
-  }
-
-  if (width * height != current_map_.data.size()) {
-    std::string message =
-      "Data size doesn't match width*height: width = " + std::to_string(width) + ", height = " +
-      std::to_string(height) + ", data size = " + std::to_string(current_map_.data.size());
-    setStatus(
-      rviz_common::properties::StatusProperty::Error, "Map", QString::fromStdString(message));
-    return;
-  }
-
-  setStatus(rviz_common::properties::StatusProperty::Ok, "Message", "Map received");
-
-  RVIZ_COMMON_LOG_DEBUG_STREAM(
-    "Received a " << current_map_.info.width << " X " <<
-      current_map_.info.height << " map @ " << current_map_.info.resolution << "m/pix\n");
-
-  showValidMap();
-}
-
-void MapDisplay::showValidMap()
-{
-  size_t width = current_map_.info.width;
-  size_t height = current_map_.info.height;
-
-  float resolution = current_map_.info.resolution;
-
-  resetSwatchesIfNecessary(width, height, resolution);
-
-  frame_ = current_map_.header.frame_id;
-  if (frame_.empty()) {
-    frame_ = "/map";
-  }
-
-  updateSwatches();
-
-  setStatus(rviz_common::properties::StatusProperty::Ok, "Map", "Map OK");
-  updatePalette();
-
-  resolution_property_->setValue(resolution);
-  width_property_->setValue(static_cast<unsigned int>(width));
-  height_property_->setValue(static_cast<unsigned int>(height));
-
-  position_property_->setVector(rviz_common::pointMsgToOgre(current_map_.info.origin.position));
-  orientation_property_->setQuaternion(
-    rviz_common::quaternionMsgToOgre(current_map_.info.origin.orientation));
-
-  transformMap();
-
-  updateDrawUnder();
-
-  context_->queueRender();
-}
-
-void MapDisplay::resetSwatchesIfNecessary(size_t width, size_t height, float resolution)
-{
-  if (width != width_ || height != height_ || resolution_ != resolution) {
-    createSwatches();
-    width_ = width;
-    height_ = height;
-    resolution_ = resolution;
-  }
-}
-
-void MapDisplay::updateSwatches() const
-{
-  for (const auto & swatch : swatches_) {
-    swatch->updateData(current_map_);
-
-    Ogre::Pass * pass = swatch->getTechniquePass();
-    Ogre::TextureUnitState * tex_unit = nullptr;
-    if (pass->getNumTextureUnitStates() > 0) {
-      tex_unit = pass->getTextureUnitState(0);
-    } else {
-      tex_unit = pass->createTextureUnitState();
-    }
-
-    tex_unit->setTextureName(swatch->getTextureName());
-    tex_unit->setTextureFiltering(Ogre::TFO_NONE);
-    swatch->setVisible(true);
-    swatch->resetOldTexture();
-  }
+  current_map_ = *msg;
+  loaded_ = true;
+  // updated via signal in case ros spinner is in a different thread
+  Q_EMIT q_helper_object_->mapUpdated();
 }
 
 void MapDisplay::updatePalette()
@@ -588,68 +171,23 @@ void MapDisplay::updatePalette()
   updateDrawUnder();
 }
 
-void MapDisplay::transformMap()
+std::shared_ptr<SwatchBase<nav_msgs::msg::OccupancyGrid>> MapDisplay::createSwatch(
+  Ogre::SceneManager * scene_manager,
+  Ogre::SceneNode * parent_scene_node,
+  size_t x,
+  size_t y,
+  size_t width,
+  size_t height,
+  float resolution,
+  bool draw_under)
 {
-  if (!loaded_) {
-    return;
-  }
-
-  rclcpp::Time transform_time = context_->getClock()->now();
-
-  if (transform_timestamp_property_->getBool()) {
-    transform_time = current_map_.header.stamp;
-  }
-
-  Ogre::Vector3 position;
-  Ogre::Quaternion orientation;
-  if (!context_->getFrameManager()->transform(
-      frame_, transform_time, current_map_.info.origin, position, orientation) &&
-    !context_->getFrameManager()->transform(
-      frame_, rclcpp::Time(0, 0, context_->getClock()->get_clock_type()),
-      current_map_.info.origin, position, orientation))
-  {
-    setMissingTransformToFixedFrame(frame_);
-    scene_node_->setVisible(false);
-  } else {
-    setTransformOk();
-
-    scene_node_->setPosition(position);
-    scene_node_->setOrientation(orientation);
-  }
-}
-
-void MapDisplay::fixedFrameChanged()
-{
-  transformMap();
-}
-
-void MapDisplay::reset()
-{
-  MFDClass::reset();
-  update_messages_received_ = 0;
-  clear();
-}
-
-void MapDisplay::update(float wall_dt, float ros_dt)
-{
-  (void) wall_dt;
-  (void) ros_dt;
-
-  transformMap();
-}
-
-void MapDisplay::onEnable()
-{
-  MFDClass::onEnable();
-  setStatus(rviz_common::properties::StatusProperty::Warn, "Message", "No map received");
-}
-
-void MapDisplay::updateMapUpdateTopic()
-{
-  unsubscribeToUpdateTopic();
-  reset();
-  subscribeToUpdateTopic();
-  context_->queueRender();
+  // make shared not possible due to abstract base class
+  return std::shared_ptr<SwatchBase<nav_msgs::msg::OccupancyGrid>>(
+    new Swatch(
+      scene_manager,
+      parent_scene_node,
+      x, y, width, height, resolution, draw_under
+  ));
 }
 
 }  // namespace displays
